Fix window flicker
Since the window positioning is calculated using the size of the window, and the size isn't known until the window has been rendered at least once (since the text length can change), the windows "flicker" as they are positioned incorrectly on the first frame and then snap into the right place on the second frame they are visible. They also do something similar any time the size of the window changes (e.g. hovering over a different item), though the bigger the change the more noticeable.

By using ImGui.SetNextWindowPos instead, we can pass in a pivot point which lets us set the position of the top right point for the "left" window instead. This gets resolved right at the end after things have mostly rendered , and there's no flickering.

I also noticed that the first window/ImGui.Begin call was missing an End call, so I added that. I also thought that there was something incorrect because the End calls were called outside of the if statements (i.e. called no matter what the return value of Begin was), which is normally not how the ImGui calls work, but looking at the [imgui header file][1], I see that ImGui.Begin (and BeginChild) is different from all the other API calls in this sense, and End should *always* be called.

[1]: https://github.com/ocornut/imgui/blob/d3bb3336f5359dd3bf4af9bb32c2164b31414eac/imgui.h#L366-L368